+++ app/Services/SurchargeService.php
@@ -8,12 +8,12 @@ use App\Models\Currency;
 class SurchargeService implements CalculatorServiceInterface
 {
     /**
-     * @param int $amount
+     * @param float $amount
      * @param Currency $currency
      * @return float
      */
-    public function calculate(int $amount, Currency $currency): float
+    public function calculate(float $amount, Currency $currency): float
     {
-        return round($amount *  ($currency->surcharge->percentage / 100), 2);
+        return round($amount * ($currency->surcharge->percentage / 100), 2);
     }
 }
